backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pyjwt
fitz
pillow
python-docx
pandas
openpyxl
xlrd
aiofiles